% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cblas_dgemv_example}
\alias{cblas_dgemv_example}
\title{Example of cblas_dgemv}
\usage{
cblas_dgemv_example(X, y, alpha = 1)
}
\arguments{
\item{X}{A matrix.}

\item{y}{A vector.}

\item{alpha}{Numerics.}

\item{z}{A vector.}
}
\value{
A numeric vector. The result of `alpha * X * y`.
}
\description{
A function to demo how to use `cblas_dgemv` which perform `z = alpha * X * y + beta * z` with `beta = 0`.
}
\examples{
X <- matrix(1:6, 3, 2)
y <- c(1, 2)
cblas_dgemv_example(X, y)
}
